// scuola/front-end/src/app/shared/interfaces.ts
export class Azienda {
    id;
    address?: string;
    description?: string;
    email?: string;
    latitude: number;
    longitude: number;
    nome?: string;
    origin?: string;
    partita_iva?: string;
    pec?: string;
    phone?: string;
    convenzione_id?: Convenzione;
    idTipoAzienda;
    coordinate;
}

export class Convenzione{
    convenzionePAT: {};
    convenzioneIstituto: {}
  }